+++ src/icons/call/PhoneCancelRecallIcon.vue
@@ -1,0 +1,10 @@
+<template>
+  <svg width="20" height="20" viewBox="0 0 20 20" fill="none" xmlns="http://www.w3.org/2000/svg">
+    <path
+      d="M7.88779 17.8855C12.2432 19.0526 16.7201 16.4679 17.8871 12.1124C19.0541 7.75699 16.4694 3.28014 12.114 2.1131C7.75856 0.946069 3.28171 3.53078 2.11468 7.88621C0.947643 12.2417 3.53235 16.7185 7.88779 17.8855Z"
+      stroke="currentColor"
+      stroke-width="2"
+    />
+    <path d="M15 15.5L5 4.5" stroke="currentColor" stroke-width="2" />
+  </svg>
+</template>