+++ src/icons/call/PhoneExecuteRecallIcon.vue
@@ -1,0 +1,5 @@
+<template>
+  <svg width="20" height="20" viewBox="0 0 17 12" fill="none" xmlns="http://www.w3.org/2000/svg">
+    <path d="M16 1L6 11L1 6" stroke="currentColor" stroke-width="2" stroke-linecap="round" stroke-linejoin="round"/>
+  </svg>
+</template>