+++ src/icons/index.ts
@@ -31,6 +31,10 @@ import ClientCallIcon from './call/ClientCallIcon.vue';
 import HangupCallIcon from './call/HangupCallIcon.vue';
 import KeyboardCallIcon from './call/KeyboardCallIcon.vue';
 import MicrophoneCallIcon from './call/MicrophoneCallIcon.vue';
+import PhoneCancelRecallIcon from './call/PhoneCancelRecallIcon.vue';
+import PhoneExecuteRecallIcon from './call/PhoneExecuteRecallIcon.vue';
+import PhoneHangupIcon from './call/PhoneHangupIcon.vue';
+import PhoneRecallIcon from './call/PhoneRecallIcon.vue';
 import ScenarioCallIcon from './call/ScenarioCallIcon.vue';
 import TransferCallIcon from './call/TransferCallIcon.vue';
 import VolumeCallIcon from './call/VolumeCallIcon.vue';
@@ -70,17 +74,19 @@ import DsrIcon from './others/DsrIcon.vue';
 import HoldIcon from './others/HoldIcon.vue';
 import IvrIcon from './others/IvrIcon.vue';
 import MinusIcon from './others/MinusIcon.vue';
-import OperatorIcon from './others/OperatorIcon.vue';
 import PlusIcon from './others/PlusIcon.vue';
 import PredictiveIcon from './others/PredictiveIcon.vue';
 import PreviewIcon from './others/PreviewIcon.vue';
 import PrioritySelectIcon from './others/PrioritySelectIcon.vue';
 import ProgressiveIcon from './others/ProgressiveIcon.vue';
 import SmsIcon from './others/SmsIcon.vue';
+import StatisticDetailIcon from './others/StatisticDetailIcon.vue';
+import StatisticSummaryIcon from './others/StatisticSummaryIcon.vue';
 import StatusSelectIcon from './others/StatusSelectIcon.vue';
 import TaskSelectIcon from './others/TaskSelectIcon.vue';
 import TranscriptionIcon from './others/TranscriptionIcon.vue';
 import TranscriptionStartIcon from './others/TranscriptionStartIcon.vue';
+import operatorIcon from './others/operatorIcon.vue';
 import PauseIcon from './player/PauseIcon.vue';
 import StartIcon from './player/StartIcon.vue';
 import StopIcon from './player/StopIcon.vue';
@@ -135,6 +141,10 @@ export const icons = {
   hangupCall: HangupCallIcon,
   keyboardCall: KeyboardCallIcon,
   microphoneCall: MicrophoneCallIcon,
+  phoneCancelRecall: PhoneCancelRecallIcon,
+  phoneExecuteRecall: PhoneExecuteRecallIcon,
+  phoneHangup: PhoneHangupIcon,
+  phoneRecall: PhoneRecallIcon,
   scenarioCall: ScenarioCallIcon,
   transferCall: TransferCallIcon,
   volumeCall: VolumeCallIcon,
@@ -184,17 +194,19 @@ export const icons = {
   hold: HoldIcon,
   ivr: IvrIcon,
   minus: MinusIcon,
-  operator: OperatorIcon,
   plus: PlusIcon,
   predictive: PredictiveIcon,
   preview: PreviewIcon,
   prioritySelect: PrioritySelectIcon,
   progressive: ProgressiveIcon,
   sms: SmsIcon,
+  statisticDetail: StatisticDetailIcon,
+  statisticSummary: StatisticSummaryIcon,
   statusSelect: StatusSelectIcon,
   taskSelect: TaskSelectIcon,
   transcription: TranscriptionIcon,
   transcriptionStart: TranscriptionStartIcon,
+  operator: operatorIcon,
 
   // player icons
   pause: PauseIcon,